requirements.txt
pydantic==1.9.0
pyyaml==6.0
kafka-python==2.0.2
influxdb-client==1.26.0
msgpack==1.0.3